Renamed EventListener to TVEventListener. Renamed ResponseListener to TVResponseListener. Renamed TouchListener to TVTouchListener. Renamed SongAdapter to MediaItemAdapter. Deleted the Song-class.
Implemented Generics to DeviceManager, DeviceCommand and DeviceObserver.
This is a MUCH better way to implement all the different types of
devices.
Have an instance of DeviceManager for each type of device, and implement
the 3 Observer-methods for each type of device. Also have a typesafe
Command-pattern for all devices.

Also refactored MediaFinder a lot. Its purpose now is a pure
helper-class to find and list songs/videos. The DeviceManager +
DeviceCommand will handle the interaction with media-items.